fix(vm): honor legacy encoding in string widths
Use the configured Emuera ANSI encoding for STRLENS and SUBSTRING so narrow Chinese FORMCELL values do not enter an invalid truncation path. Add a regression for the eraRorona failure and align existing string-width expectations.

+++ crates/era-runtime/src/session/debug_session.rs
@@ -1267,7 +1267,7 @@ mod console_tests {
         );
         assert_eq!(
             parse_console_expression("STRLENS(\"界\")", &[]),
-            Ok(VmValue::Integer(3))
+            Ok(VmValue::Integer(2))
         );
         assert_eq!(
             parse_console_expression("STRLENSU(\"😀\")", &[]),
